merge

+++ Dimitri/Game/Game/src/Models/Buttons/ExitButton.cpp
@@ -3,7 +3,7 @@
 
 namespace Game {
 	namespace Models {
-		Buttons::ExitButton::ExitButton(int x, int y, int z, int height, int width, Enums::DirectionEnum state, int scene_height, Graphics::Models::Center center) :
+		Buttons::ExitButton::ExitButton(float x, float y, float z, float height, float width, Enums::DirectionEnum state, float scene_height, Graphics::Models::Center center) :
 			Button(x, y, z, height, width, state, scene_height, center, "ExitButton", Game::Enums::ButtonEnum::EXIT) {
 			initialize_textures();
 		}
